Fix mixing potion

+++ lib/other/constants.dart
@@ -5,4 +5,4 @@ enum PotionState { empty, mixing, finished }
 const int potionShakeIntervalMs = 2000;
 const int potionShakeThreshold = 10;
 const int maxMixLevel = 10;
-const double mixLevelIncrease = 1.0;
+const int mixLevelIncrease = 1;
